// frontend/src/api/generated/model/restaurant.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Delivery API
 * API para gestión de restaurantes
 * OpenAPI spec version: 1.0.0
 */
import type { MenuCategory } from './menuCategory';
import type { Dish } from './dish';

export interface Restaurant {
  id: number;
  name: string;
  description?: string;
  cuisine?: string;
  /**
   * @minimum 0
   * @maximum 5
   */
  rating?: number;
  reviewCount?: number;
  deliveryTimeMin?: number;
  deliveryTimeMax?: number;
  deliveryFee?: number;
  minimumOrder?: number;
  isOpen: boolean;
  imageUrl?: string;
  categories?: MenuCategory[];
  dishes?: Dish[];
}